ConTaxi Commands & CenTaxi choose taxi

+++ CenDLL/DLLMap.h
@@ -2,9 +2,9 @@
 #include "CenDLL.h"
 
 typedef struct XYObject{
-	float coordX;	//Horizontal position
-	float coordY; //Vertical position
-	float speedX; //1 = (Moving Right) | -1 = (Moving Left) | 0 = (Horizontally Stationary)
-	float speedY; //1 = (Moving Down) | -1 = (Moving Up) | 0 = (Vertically Stationary)
-	float speedMultiplier; //Speed * Multiplier
+	double coordX;	//Horizontal position
+	double coordY; //Vertical position
+	double speedX; //1 = (Moving Right) | -1 = (Moving Left) | 0 = (Horizontally Stationary)
+	double speedY; //1 = (Moving Down) | -1 = (Moving Up) | 0 = (Vertically Stationary)
+	double speedMultiplier; //Speed * Multiplier
 }XYObject;
